requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Django==2.1.4
django-appconf==1.0.2
django-autocomplete-light==3.3.2
django-geojson==2.12.0
django-leaflet==0.24.0
django-settings-export==1.2.1
django-suit==0.2.26
idna==2.7
isort==4.3.4
Jinja2==2.10
jsonfield==2.0.2
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
packaging==18.0
pep8==1.7.1
Pillow==5.3.0
pip-autoremove==0.9.1
pkg-resources==0.0.0
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pylint==2.2.2
pylint-django==2.0.4
pylint-plugin-utils==0.4
pyparsing==2.3.0
pytz==2018.7
reportlab==3.5.12
requests==2.20.1
rope==0.11.0
six==1.11.0
typed-ast==1.1.0
tzlocal==1.5.1
urllib3==1.24.1
virtualenv==16.1.0
wrapt==1.10.11
